Add process info printing
Calling ps now prints out a header, followed by the info of the currently active processes.
The process info includes the pid, parent pid, state, size, and name.
Each line refers to its own process.
When it finishes, ps returns the number of active processes, or -1 if there's an error.